- Controller objects modified to support (for test) the headV3 boards (b12,b13). For this test all the joints are supposed to be indipendent, we'll modify the file adding a joints coupling later on - modified the calibration5 parameters to be int32_t values

+++ emBODY/eBcode/arch-arm/board/mc4plus/appl/wip/src/eoappservices/mc/EOaxisController.c
@@ -159,7 +159,7 @@ extern void eo_axisController_StartCalibration_type3(EOaxisController *o)
     o->control_mode = eomc_controlmode_calib;
 }
 
-extern void eo_axisController_StartCalibration_type5(EOaxisController *o, int16_t pwmlimit, int16_t vel, int32_t final_position)
+extern void eo_axisController_StartCalibration_type5(EOaxisController *o, int32_t pwmlimit, int32_t vel, int32_t final_position)
 {
     if (!o) return;
